requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
python-dotenv
faiss-cpu
huggingface-hub
InstructorEmbedding
sentence-transformers
pandas
